+++ core-ng/src/main/java/core/framework/inject/Inject.java
@@ -4,13 +4,12 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
 import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /**
  * @author neo
  */
-@Target({FIELD, METHOD})
+@Target(FIELD)
 @Retention(RUNTIME)
 public @interface Inject {
 }