+++ core-ng/src/main/java/core/framework/inject/Named.java
@@ -4,13 +4,12 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
 import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /**
  * @author neo
  */
-@Target({FIELD, PARAMETER})
+@Target(FIELD)
 @Retention(RUNTIME)
 public @interface Named {
     String value();