+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 7.4.10 (06/10/2020 - )
 * monitor: tweak kube monitor pod not ready error message
 * stats: track jvm_non_heap_used by default
+* cache: gracefully handle incompatible json structure from redis, to avoid clearing cache keys manually
 
 ### 7.4.9 (06/04/2020 - 06/09/2020)
 * http: for static content controllers, convert client abort exception to warning (still better use CDN/Storage solution to serve static resources)